Still getting lock errors

+++ micromamba/src/run.cpp
@@ -161,9 +161,8 @@ namespace mamba
                        std::unique_ptr<LockFile> proc_dir_lock = lock_proc_dir())
             : location{ proc_dir() / fmt::format("{}.json", getpid()) }
         {
-            // Uncomment if we don't want to create json lockfiles
-            // if (Context::instance().disable_lockfile)
-            //     return;
+            if (Context::instance().disable_lockfile)
+                return;
             assert(proc_dir_lock);  // Lock must be hold for the duraction of this constructor.
 
             const auto open_mode = std::ios::binary | std::ios::trunc | std::ios::out;
@@ -184,6 +183,8 @@ namespace mamba
 
         ~ScopedProcFile()
         {
+            if (Context::instance().disable_lockfile)
+                return;
             const auto lock = lock_proc_dir();
             std::error_code errcode;
             const bool is_removed = fs::remove(location, errcode);
